% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{accumulate}
\alias{accumulate}
\title{accumulate}
\usage{
accumulate(n, ri, ro, ia, it = 100)
}
\arguments{
\item{n}{is the number of accumulated arrivals}

\item{ri}{is the rate in, a float}

\item{ro}{is the rate out, a float}

\item{ia}{is a vector of inherited ages, integers}

\item{it}{is the number of iterations of the bootstrap, an integer}
}
\value{
the median deposit ages of `n` accumulations after `it` bootstraps
}
\description{
bootstrap wrapper for recorder function
}
