% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{accumulate1}
\alias{accumulate1}
\title{accumulate1}
\usage{
accumulate1(n, ri, ro, ia, mn, it = 100)
}
\arguments{
\item{n}{is the number of accumulated arrivals}

\item{ri}{is the rate in, a float}

\item{ro}{is the rate out, a float}

\item{ia}{is a vector of inherited ages, integers}

\item{mn}{is the minimum age}

\item{it}{is the number of iterations of the bootstrap, an integer}
}
\value{
the median deposit ages of `n` accumulations after `it` bootstraps
}
\description{
bootstrap wrapper for recorder function
adds minimum age
}
\seealso{
recorder1
}
