% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{accumulater1}
\alias{accumulater1}
\title{accumulater1}
\usage{
accumulater1(n, ti_rng, to_rng, ia, obs, batch = 10, it = 100)
}
\arguments{
\item{n}{is the number of accumulated arrivals}

\item{ti_rng}{is the range of rates in, a vector of floats}

\item{to_rng}{is the range of rates out, a vector of floats}

\item{ia}{is a vector of inherited ages, integers}

\item{obs}{is the vector of observed estimated mean charcoal ages}

\item{batch}{is the batch size of random draws}

\item{it}{is the number of iterations of the bootstrap, an integer}
}
\value{
the median deposit ages of `n` accumulations after `it` bootstraps
}
\description{
bootstrap wrapper for accumulate function
adds minimum age
}
\seealso{
accumulate

recorder
}
