% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{ages}
\alias{ages}
\title{ages}
\usage{
ages(vals, index)
}
\arguments{
\item{vals}{is a numeric vector of variable values}

\item{index}{is a reference index of values inclusive of `vals`}
}
\value{
vector of `vals` rounded to nearest `index` value
}
\description{
ages
}
