% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{arrivals}
\alias{arrivals}
\title{arrivals}
\usage{
arrivals(vals, index)
}
\arguments{
\item{vals}{is a numeric vector of variable values}

\item{index}{is a reference index of values inclusive of `vals`}
}
\value{
a vector length `index` showing the number of `vals` in each rounding to the nearest index
}
\description{
arrivals
}
