% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{avg_slope}
\alias{avg_slope}
\title{avg slope}
\usage{
avg_slope(el, hip)
}
\description{
naive average slope function
}
\details{
@param el is a vector of elevations
 @param hip is a vector of hipchain lengths (distance to outlet)
 @return local average bed slope
 @export
}
