% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{back_window}
\alias{back_window}
\title{back_window}
\usage{
back_window(vals, index, window = 500, type = "float")
}
\arguments{
\item{vals}{is a numeric vector of values}

\item{index}{is a numeric vector of index values}

\item{window}{is a number specifying the length of background window}

\item{type}{is a string, options = c('float', 'discrete', 'log', 'float_log')}
}
\value{
list of `vals` in `window` along `index`
}
\description{
divides values associated with index numbers into subsets based upon a background window
}
