% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{backfill1}
\alias{backfill1}
\title{backfill1}
\usage{
backfill1(nodes, n, vi, vo, tp, lp, it = 10)
}
\arguments{
\item{nodes}{is a spatial object (creek nodes)}

\item{n}{is the number of years to simulate accumulation, an integer}

\item{vi}{is the volumetric input rate}

\item{vo}{is the volumetric output rate}

\item{tp}{is the turbulent deposition probability}

\item{lp}{is the laminar deposition probability}

\item{it}{is the number of iterations to simulate accumulation, an integer}
}
\value{
the median accumulation record after `it` iterations in volume and level
}
\description{
Function to fit volume rate to observed volumes.
Creek nodes are linked, passing deposits downstream.
Deposits backfill based upon average elevation
}
\seealso{
record_volume
}
