% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin_by}
\alias{bin_by}
\title{bin by}
\usage{
bin_by(vec, var, bins = 10)
}
\arguments{
\item{vec}{is the vector to divide by}

\item{var}{is the variable to sum}

\item{bins}{is the number of bins to divide into}
}
\value{
a plot and dataframe of mean `var` per bin
}
\description{
Divides a vector by length into bins, takes the sum of a corresponding variable.
}
