% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{bin_cdfs}
\alias{bin_cdfs}
\title{bin_cdfs}
\usage{
bin_cdfs(num, den, by, rho = 1, bins = 10)
}
\arguments{
\item{num}{is a numeric vector}

\item{den}{is a numeric vector}

\item{by}{is a numeric vector}

\item{rho}{is a number representing density}

\item{bins}{is the number of bins to use}
}
\value{
a dataframe of change in cdf of each var by bin
}
\description{
returns a dataframe of change in cdf of each var by bin
}
