% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{bin_weights}
\alias{bin_weights}
\title{bin_weights}
\usage{
bin_weights(num, den, by, bins = 10)
}
\arguments{
\item{num}{is a numeric vector}

\item{den}{is a numeric vector}

\item{by}{is a numeric vector}

\item{bins}{is the number of bins to use}
}
\value{
a dataframe of change in cdf of each var by bin
}
\description{
returns a dataframe of change in cdf of each var by bin
}
