% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{birdseye}
\alias{birdseye}
\title{birdseye}
\usage{
birdseye(pt, mag = 40, flow = flowline, nod = nodes, map = uk)
}
\description{
convenience plotting function
}
\details{
@param pt is a coordinate pair (x,y)
 @param mag is an integer of magnification
 @param flow is the river network flowline
 @param nod is the river node output of the M/B model
 @param map is the background raster
 @return a plot centered on `pt` at zoom `mag`
 @importFrom magrittr %>%
 @seealso circ
 @export
}
