% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{boot_accum}
\alias{boot_accum}
\title{boot_accum}
\usage{
boot_accum(n, ti_rng, to_rng, ia, obs, batch = 1000, it = 100)
}
\arguments{
\item{n}{is the number of accumulated arrivals}

\item{ti_rng}{is the range of rates in, a vector of floats}

\item{to_rng}{is the range of rates out, a vector of floats}

\item{ia}{is a vector of inherited ages, integers}

\item{obs}{is the vector of observed estimated mean charcoal ages}

\item{batch}{is the batch size of random draws}

\item{it}{is the number of iterations of the bootstrap, an integer}
}
\value{
the result with the lowest ks test value
}
\description{
bootstrap utility for estimating confidence intervals
wraps accumulater()
}
\seealso{
accumulater

accumulate

recorder
}
