% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boot_ids}
\alias{boot_ids}
\title{boot ids}
\usage{
boot_ids(n, rl)
}
\arguments{
\item{n}{is an integer specifying the number of bootstraps}

\item{rl}{is a rank list output from function `rank_list`}
}
\value{
`n` bootstraps with replacement of length `rl`
}
\description{
wrapper for lapply, given a rank list from a site, returns
`n` bootstraps with replacement of equal length as the rank list
}
\seealso{
rank_list
}
