% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cdf_dif}
\alias{cdf_dif}
\title{cdf_dif}
\usage{
cdf_dif(cdf)
}
\arguments{
\item{cdf}{is a numeric vector increasing monotonically between zero and 1}
}
\value{
a vector of differences between increments of `cdf`
}
\description{
returns a vector of differences between increments of `cdf`
}
