% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{cdf_gof}
\alias{cdf_gof}
\title{cdf_gof}
\usage{
cdf_gof(ob1, ob2)
}
\arguments{
\item{ob1}{is a cdf, a numeric vector}

\item{ob2}{is a cdf, a numeric vector}
}
\value{
the k-s stat and kuiper stat
}
\description{
compares two cdfs for goodness of fit, returns K-S stat and kuiper stat
}
