% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{circ}
\alias{circ}
\title{draw circle}
\usage{
circ(pt, rad = 15)
}
\arguments{
\item{pt}{is a coordinate pair}

\item{rad}{is the numeric radius of the circle}
}
\value{
matrix of coordinates for a circle of radius `rad` with center `pt`
}
\description{
draw circle
}
\seealso{
tri_length
}
