% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cis_of_boot}
\alias{cis_of_boot}
\title{CIs of Boot}
\usage{
cis_of_boot(boot, pmfs = char_pmfs)
}
\arguments{
\item{boot}{is a matrix of pmfs}

\item{pmfs}{is the reference matrix of pmfs (charcoal pmfs)}
}
\value{
vector of cis of `boot`
}
\description{
Return confidence intervals from a matrix of pmf bootstraps.
}
