% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{contr_delta}
\alias{contr_delta}
\title{contr delta}
\usage{
contr_delta(vec)
}
\arguments{
\item{vec}{is a numeric vector of contributing areas}
}
\value{
a vector of differences in contributing area per step
}
\description{
utility function for identifying tributaries
}
