% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inher.R
\name{convo_lis}
\alias{convo_lis}
\title{convo_lis subtracts the first distribution from each distribution in a list.}
\usage{
convo_lis(lis, dat, vec)
}
\arguments{
\item{lis}{is a list of numeric sorted ranks}

\item{dat}{is a matrix of pdfs rows with cols of obs sorted by rank}

\item{vec}{is a numeric vector of years}
}
\value{
a matrix of pdfs convolved by subtracting pdf of lowest rank from pdfs
  with ranks in lis
}
\description{
Given a list of numeric sorted ranks (lis), a matrix of pdf rows with
cols of obs sorted by rank (dat), and a numeric vector of years (vec),
returns a matrix of pdfs convolved by subtracting the pdf in dat of lowest rank
in lis with the set of pdfs with ranks in lis.
}
