% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convo_list}
\alias{convo_list}
\title{convo list}
\usage{
convo_list(ranks, pmfs = char_pmfs)
}
\arguments{
\item{ranks}{is a vector of ranks or list of rank vectors}

\item{pmfs}{is a matrix with cols obs and rows probs}
}
\value{
convolved pmfs by subtraction
}
\description{
convolves pmfs by rank list
}
