% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convo_rank}
\alias{convo_rank}
\title{convolve by rank}
\usage{
convo_rank(ranks, pmfs = char_pmfs, vals = as.numeric(colnames(pmfs)),
  dat = charcoal)
}
\arguments{
\item{ranks}{is a vector of ranks (element of rank list)}

\item{pmfs}{is the matrix of charcoal pmfs}

\item{vals}{is the vector of values (years) associated with pmfs}

\item{dat}{is the charcoal variable data.table}
}
\value{
pmfs of ranks convolved by subtracting youngest from each
}
\description{
Given a vector of ranks, pulls the pmfs for each rank and subtracts
the youngest from each via convolution.
}
