% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{cor_creeks}
\alias{cor_creeks}
\title{cor_creeks}
\usage{
cor_creeks(
  var1,
  var2,
  creeks,
  creek_labs,
  labs = c("var1", "var2"),
  cols = length(creek_labs),
  leg_pos = "topleft",
  a = 0.33
)
}
\arguments{
\item{var1}{is a numeric vector}

\item{var2}{is a numeric vector length(`var1`)}

\item{creeks}{is a character vector of creek names}

\item{creek_labs}{is a character vector providing creek labels}

\item{labs}{is a character vector providing two plot labels in order `var1` `var2`}

\item{leg_pos}{is a character vector denoting legend position}

\item{a}{is the level of alpha transparency (0,1)}
}
\value{
a scatter plot of `var1` vs. `var2` with a linear fit line
}
\description{
prints a correlation plot of two variables colored by creek
}
