% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{cor_plot}
\alias{cor_plot}
\title{cor_plot}
\usage{
cor_plot(var1, var2, labs = c("var1", "var2"))
}
\arguments{
\item{var1}{is a numeric vector}

\item{var2}{is a numeric vector length(`var1`)}

\item{labs}{is a character vector providing two plot labels in order `var1` `var2`}
}
\value{
a scatter plot of `var1` vs. `var2` with a linear fit line
}
\description{
prints a correlation plot of two variables
}
