% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{differ}
\alias{differ}
\title{differ}
\usage{
differ(vec, delta = 0)
}
\arguments{
\item{vec}{is a numeric vector of length > 1}
}
\value{
a vector of differences between elements in `vec`
}
\description{
given a vector of length > 1, returns a vector differences between elements
}
