% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{discrete_bin}
\alias{discrete_bin}
\title{discrete_bin}
\usage{
discrete_bin(vals, bin_len)
}
\arguments{
\item{vals}{is a numeric vector, the variable of interest}

\item{bin_len}{is the interval to divide `vals` by}
}
\value{
a list of ids for `vals` divided by `bin_len` into bins
}
\description{
divide a variable of interest `vals`
into intervals of length `bin_len`
return ids of `vals` as a list
}
