% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distribute}
\alias{distribute}
\title{distribute}
\usage{
distribute(pdf, val, size = 10000)
}
\arguments{
\item{pdf}{is a vector of probs summing to one}

\item{val}{is a vector of values associated with `pdf`}

\item{size}{is a large integer (>100)}
}
\value{
sample of vals reflective of pdf
}
\description{
given a pdf for a vector of values, returns a sample space of vals distributed
around the pdf
}
