% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_exp_decay}
\alias{draw_exp_decay}
\title{Draw from exponential decay distribution}
\usage{
draw_exp_decay(n, k)
}
\arguments{
\item{n}{is the max value returned by the decay function (integer)}

\item{k}{is the exponent $x^k$ (numeric)}
}
\value{
\code{n} draws from the exponential decay distribution.
}
\description{
Draw from exponential decay distribution $y = x^k$
}
