% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{emp_cdf}
\alias{emp_cdf}
\title{emp_cdf}
\usage{
emp_cdf(vals)
}
\arguments{
\item{vals}{is a numeric vector of variable values}
}
\value{
the empiric cdf of `vals`
}
\description{
derives the empiric cdf of a numeric vector of values
}
