% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{eventer}
\alias{eventer}
\title{eventer}
\usage{
eventer(rt, id, vl, n, lab)
}
\arguments{
\item{rt}{is the temporal event rate}

\item{id}{is the creek node number, an integer.}

\item{vl}{is the volumetric event rate}

\item{n}{is the period of time to simulate events}

\item{lab}{is the type label, which sounds like a string but is numeric!}
}
\value{
a data.frame of events with cols (age, vol, type, id)
}
\description{
Utility for voluminous().
Generates an event data.frame with cols (age, vol, type, id),
given a temporal and volumetric event rate, over a given period of time.
}
