% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{events}
\alias{events}
\title{events}
\usage{
events(vals, index)
}
\arguments{
\item{vals}{is a numeric vector of variable values}

\item{index}{is an numeric vector of index values for `vals`}
}
\value{
a count of vals rounding to values of `index`
}
\description{
round a series of values to nearest index number in counts
}
