% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluvial.R
\name{fines_fit_n}
\alias{fines_fit_n}
\title{fines_fit_n}
\usage{
fines_fit_n(
  n,
  capture,
  storage,
  turnover,
  storage_index,
  source_index,
  source_prob,
  gravel_index,
  gravel_prob,
  fines
)
}
\arguments{
\item{n}{The number of runs to simulate.}

\item{capture}{The capture rate, ranging 0-1.}

\item{storage}{The storage rate, ranging 0-1.}

\item{turnover}{The turnover period, in years.}

\item{storage_index}{is the age value index for storage}

\item{source_index}{is the age value index corresponding to the PMF source_prob}

\item{source_prob}{is the PMF of debris-flow deposit ages}

\item{gravel_index}{is the age value index corresponding to the PMF gravel_prob}

\item{gravel_prob}{is the PMF of gravel ages}

\item{fines}{A vector of observed gravel ages.}
}
\value{
mean hit % of test thresholds, using test_fit()
}
\description{
Fit multiple runs of fines_synth() to observed deposits.
}
\seealso{
fines_synth

test_fit
}
