% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluvial.R
\name{fines_synth}
\alias{fines_synth}
\title{fines_synth}
\usage{
fines_synth(
  capture,
  storage,
  turnover,
  storage_index,
  source_index,
  source_prob,
  gravel_index,
  gravel_prob
)
}
\arguments{
\item{capture}{the capture rate 0-1}

\item{storage}{the storage rate 0-1}

\item{turnover}{the turnover period}

\item{storage_index}{is the age value index for storage}

\item{source_index}{is the age value index corresponding to the PMF source_prob}

\item{source_prob}{is the PMF of debris-flow deposit ages}

\item{gravel_index}{is the age value index corresponding to the PMF gravel_prob}

\item{gravel_prob}{is the PMF of gravel ages}
}
\value{
vector of synthetic fines ages
}
\description{
simulate fines ages
}
