% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluvial.R
\name{fish}
\alias{fish}
\title{fish}
\usage{
fish(rt, t, k)
}
\arguments{
\item{rt}{is the rate.}

\item{t}{is the time interval.}

\item{k}{is the number of events.}
}
\value{
a p(k) Poisson distribution, the probability of k events in time t.
}
\description{
Given a rate, time and number of events,
returns a p(k) Poisson distribution.
}
