% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{fit_bins}
\alias{fit_bins}
\title{fit_bins}
\usage{
fit_bins(dat, ids, bins, dist = "exp", index = as.numeric(rownames(char_pmfs)))
}
\arguments{
\item{dat}{is a numeric matrix of pmfs (pmf, samples)}

\item{ids}{is a numeric vector of expected means of samples in `dat`}

\item{bins}{is a list of numeric matrices of pmfs (pmf, samples)}

\item{dist}{is the distribution used by `fitdistrplus::fitdist`}

\item{index}{is a numeric vector of values associated with pmf in `dat`}
}
\value{
a list of interarrival times, exp fits, med and ci values
}
\description{
for each bin in `bins` select pmfs in `dat`
where expected means in bin equals `ids` in `dat`
compute interarrival times
as expected mean of convolved difference of successive arrivals
fit to exponential distribution
return interarrival times, exp fit, med and 95% cis
}
