% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inher.R
\name{fit_pdf}
\alias{fit_pdf}
\title{FIT PDF TO VALUE VECTOR}
\usage{
fit_pdf(vals, index)
}
\arguments{
\item{vals}{is a numeric vector of values}

\item{index}{is a numeric index of values}
}
\value{
a numeric vector containing the pdf of \code{vals} along \code{index}
}
\description{
Given a numeric vector of values \code{vals}, and an index of values \code{index},
find the pdf of \code{vals}
}
