% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inher.R
\name{fit_pmf}
\alias{fit_pmf}
\title{Find the PMF of values vector.}
\usage{
fit_pmf(vals, index)
}
\arguments{
\item{vals}{is a numeric vector of values}

\item{index}{is a numeric index of values}
}
\value{
a numeric vector containing the pmf of \code{vals} along \code{index}
}
\description{
Given a numeric vector of values \code{vals}, and an index of values \code{index},
find the pmf of \code{vals}
}
