% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{fit_volumes}
\alias{fit_volumes}
\title{fit_volumes}
\usage{
fit_volumes(nodes, n, vi, vo, it = 10)
}
\arguments{
\item{nodes}{is a spatial object (creek nodes)}

\item{n}{is an integer representing years to simulate accumulation}

\item{vi}{is the volumetric input rate, a float}

\item{vo}{is the volumetric output rate, a float}

\item{it}{is an integer representing the number of iterations per node}
}
\value{
a data.frame with cols c('vol', 'lvl')
}
\description{
function to fit volumes to creek nodes
}
