% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{fit_volumes1}
\alias{fit_volumes1}
\title{fit_volumes1}
\usage{
fit_volumes1(nodes, n, vi, vo, it = 10)
}
\arguments{
\item{nodes}{is a spatial object (creek nodes)}

\item{n}{is the number of years to simulate accumulation, an integer}

\item{vi}{is the volumetric input rate}

\item{vo}{is the volumetric output rate}

\item{it}{is the number of iterations to simulate accumulation, an integer}
}
\value{
the median accumulation record after `it` iterations in volume and level
}
\description{
Function to fit volume rate to observed volumes.
Creek nodes are linked, passing deposits downstream.
}
\seealso{
record_volume
}
