% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluvial.R
\name{fluvial_fit}
\alias{fluvial_fit}
\title{fluvial_fit}
\usage{
fluvial_fit(
  batch = 10,
  n = 10,
  min_cap = 0,
  max_cap = 1,
  min_stor = 0,
  max_stor = 1,
  min_turn = 50,
  max_turn = 10000,
  storage_index,
  source_index,
  source_prob,
  fines = F
)
}
\arguments{
\item{batch}{The number of times to randomly select model parameters.}

\item{n}{The number of runs to simulate at selected paramaters.}

\item{min_cap}{= The minimum capture rate to consider, ranging 0-1, less than max_cap.}

\item{max_cap}{= The maximum capture rate to consider, ranging 0-1, greater than min_cap.}

\item{min_stor}{= The minimum storage rate to consider, ranging 0-1, less than max_stor.}

\item{max_stor}{= The maximum storage rate to consider, ranging 0-1, greater than min_stor.}

\item{min_turn}{= The minimum turnover period to consider, in years, less than max_turn.}

\item{max_turn}{= The maximum turnover period to consider, in years, greater than min_turn.}

\item{storage_index}{is the age value index for storage}

\item{source_index}{is the age value index corresponding to the PMF source_prob}

\item{source_prob}{is the PMF of debris-flow deposit ages}

\item{fines}{A vector of observed gravel ages.}
}
\value{
table of mean hit % of test thresholds, using test_fit()
}
\description{
Fit multiple fluvial deposit runs to observed deposits.
}
\seealso{
fines_synth_n

gravel_synth_n

test_fit
}
