% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{focus}
\alias{focus}
\title{focus}
\usage{
focus(test, df)
}
\arguments{
\item{test}{is a logical test length of nrow(`df`)}

\item{df}{is a dataframe}
}
\value{
first and second columns min and max
}
\description{
frame extent of data via logical test of index
}
