% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{get_bear}
\alias{get_bear}
\title{get bearing}
\usage{
get_bear(coords)
}
\arguments{
\item{coords}{is a matrix of two coordinate pairs with rows (x,y)}
}
\value{
bearing of angle of line crossing through `coords` in degrees
}
\description{
given a matrix of two coordinate pairs with rows (x,y), returns the
bearing of the angle of the line cross through `coords` in degrees
}
