% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inher.R
\name{get_cis}
\alias{get_cis}
\title{get_cis returns confidence intervals of a vector}
\usage{
get_cis(x, lwr = 0.025, mid = 0.5, upr = 0.975)
}
\arguments{
\item{x}{is a numeric vector of probabilities}
}
\value{
a numeric vector with values at 2.5\%, median and 97.5\% of the distribution (x)
}
\description{
Given a vector of probabilities (x), returns a numeric vector with
values at 2.5\%, median and 97.5\% of the distribution (x)
}
