% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{get_form}
\alias{get_form}
\title{get_form}
\usage{
get_form(df)
}
\arguments{
\item{df}{is a data.frame of data to model}
}
\value{
a formula made from colnames of `df`
}
\description{
from colnames of `df`, make into lm formula
first var is dependent, other vars independent
}
