% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{get_forms}
\alias{get_forms}
\title{get_forms}
\usage{
get_forms(df)
}
\arguments{
\item{df}{is a data.frame object}
}
\value{
character vector of formula combinations
}
\description{
from the colnames in `df`
first col is depvar, others indepvars
return character vector of all unique formulas
for feeding to lm() or glm()
}
