% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluvial.R
\name{gof}
\alias{gof}
\title{gof}
\usage{
gof(synth, obs)
}
\arguments{
\item{synth}{Vector of snythetic deposit ages.}

\item{obs}{Vector of observed deposit ages.}
}
\value{
Vector of fits = c(Anderson-Darling, Chi-Squared, Kuiper, Kolmogorov-Smirnov)
}
\description{
Goodness-of-fit statistics for fluvial deposits.
}
