% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{gof_tab}
\alias{gof_tab}
\title{gof_tab}
\usage{
gof_tab(fits, dists)
}
\arguments{
\item{fits}{is a list of lists (dist type(bins))}

\item{dists}{is a character vector of dist type labels}
}
\value{
a table of gof stats
}
\description{
`fits` is a list of list (dist type(bins))
with elements of bins holding output from `fitdistrplus::fitdist`
fit using distribution dist type
`dists` specifies labels for distribution types for gof table
table of gof stats include bin no, dist type, aic, bic, ks, ad, cvm
}
