% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluvial.R
\name{gravel_fit_n}
\alias{gravel_fit_n}
\title{gravel_fit_n}
\usage{
gravel_fit_n(
  n,
  capture,
  storage,
  turnover,
  storage_index,
  source_index,
  source_prob,
  gravels
)
}
\arguments{
\item{n}{The number of runs to simulate.}

\item{capture}{The capture rate, ranging 0-1.}

\item{storage}{The storage rate, ranging 0-1.}

\item{turnover}{The turnover period, in years.}

\item{storage_index}{is the age value index for storage}

\item{source_index}{is the age value index corresponding to the PMF source_prob}

\item{source_prob}{is the PMF of debris-flow deposit ages}

\item{gravels}{A vector of observed gravel ages.}
}
\value{
mean hit % of test thresholds, using test_fit()
}
\description{
Fit multiple runs of gravel_synth() to observed deposits.
}
\seealso{
gravel_synth

test_fit
}
