% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluvial.R
\name{gravel_synth}
\alias{gravel_synth}
\title{gravel_synth}
\usage{
gravel_synth(
  capture,
  storage,
  turnover,
  storage_index,
  source_index,
  source_prob
)
}
\arguments{
\item{capture}{the capture rate 0-1}

\item{storage}{the storage rate 0-1}

\item{turnover}{the turnover period}

\item{storage_index}{is the age value index for storage}

\item{source_index}{is the age value index corresponding to the PMF source_prob}

\item{source_prob}{is the PMF of debris-flow deposit ages}
}
\value{
vector of synthetic fines ages
}
\description{
simulate gravels ages
}
