% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{hip_to_out}
\alias{hip_to_out}
\title{hipchain to outlet}
\usage{
hip_to_out(hipchain, trib_hip, trib_out, out_hip)
}
\arguments{
\item{hipchain}{is a numeric vector of hipchain distances}

\item{trib_hip}{is a numeric vector of tributary hipchain distances}

\item{trib_out}{is a numeric vector of tributary outlet distances}

\item{out_hip}{is a numeric vector of outlet to hipchain ratios}
}
\value{
outlet distances converted from hipchain distances
}
\description{
converts hipchain lengths to outlet distances given the outlet to hipchain ratio
}
