% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{inc_path}
\alias{inc_path}
\title{increment pathway}
\usage{
inc_path(coords)
}
\arguments{
\item{coords}{is a matrix of (x,y) coordinate pairs}
}
\value{
line path incremented into segments appr unit value in length
}
\description{
given a matrix of coordinates represented a line, returns a line path
incremented into segments of approximately unit length.
not for use on branching line paths.
given a matrix of coordinates representing a path or line
take the bearing and length of each segment
divide the segment into length-1 points
for each point along the segment, calculate the x and y position
return a new path consisting containing all old and new x and y values
the purpose of this function is to increment along the channel line
so I can take more transects and get a finer resolution of detail on the channel
}
\seealso{
tri_length
}
