% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{interp_by_node}
\alias{interp_by_node}
\title{interpolate by node}
\usage{
interp_by_node(vals, trib_out, trib_id, chan_out, chan_id)
}
\arguments{
\item{vals}{is a vector of values at tributary nodes}

\item{trib_out}{is a vector of tributary outlet distances}

\item{trib_id}{is a vector of tributary node ids}

\item{chan_out}{is a vector of channel outlet distances}

\item{chan_id}{is a vector of channel node ids}
}
\value{
interpolated `vals` based on outlet distances of tribs
}
\description{
interpolate values based on outlet distances of nodes
}
