% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{logit_stat}
\alias{logit_stat}
\title{logit_stat}
\usage{
logit_stat(df)
}
\arguments{
\item{df}{is a data.frame of data to model}
}
\value{
a glm model summary fit to `df` using family binomial
}
\description{
print logit model summary given data frame
first var is depedent, other vars independent
}
