% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{low_path}
\alias{low_path}
\title{low path}
\usage{
low_path(coords, rad = 25, map = uk)
}
\arguments{
\item{coords}{is a matrix of (x,y) coordinate pairs}

\item{rad}{is a radius in meters}

\item{map}{is elevation data (DEM or lidar)}
}
\value{
the coordinates of the low path within radius of coords
}
\description{
Given a matrix of coordinates representing a line, takes transects of
radius `rad` at a bearing perpendicular to the angle of the line,
returns the low point along each transect in a matrix representing
the low path.
}
\seealso{
inc_path

take_transect
}
