% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{mod_stat}
\alias{mod_stat}
\title{mod_stat}
\usage{
mod_stat(df)
}
\arguments{
\item{df}{is a data.frame of data to model}
}
\value{
a lm model summary fit to `df`
}
\description{
print lm model summary given data frame
first var is depedent, other vars independent
}
