% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{n_dimensional_search}
\alias{n_dimensional_search}
\title{n_dimensional_search}
\usage{
n_dimensional_search(
  so,
  ti_rng,
  to_rng,
  vi_rng,
  vo_rng,
  ti_scl = 1,
  to_scl = 1,
  ks_scl = 1,
  dp_scl = 1,
  m_scl = 1,
  b_scl = 1,
  type = 0,
  it = 100,
  n = 10000,
  batch = 10
)
}
\arguments{
\item{so}{is a spatial object (creek nodes)}

\item{ti_rng}{is the range of temporal input rates to search}

\item{to_rng}{is the range of temporal output rates to search}

\item{vi_rng}{is the range of volumetric input rates to search}

\item{vo_rng}{is the range of volumetric input rates to search}

\item{ti_scl}{is the range of temporal input scales to search}

\item{to_scl}{is the range of temporal output scales to search}

\item{ks_scl}{is the range of streampower scales to search}

\item{dp_scl}{is the range of deposit probability scales to search}

\item{b_scl}{is the proportion of volume that settles out of deposits at obstructions}

\item{it}{is the number of iterations to perform at each point}

\item{n}{is the number of years to simulate accumulation, an integer}
}
\value{
a list with elements (record, mean output rate)
}
\description{
For finding a needle in a haystack.
Uses backfill() to simulate accumulation.
}
\seealso{
backfill
}
