% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{ndim_search}
\alias{ndim_search}
\title{ndim_search}
\usage{
ndim_search(
  so,
  ti_rng,
  to_rng,
  vi_rng,
  vo_rng,
  tp = 0.1,
  lp = 0.5,
  it = 100,
  n = 10000,
  batch = 10
)
}
\arguments{
\item{so}{is a spatial object (creek nodes)}

\item{ti_rng}{is the range of temporal input rates to search}

\item{to_rng}{is the range of temporal output rates to search}

\item{vi_rng}{is the range of volumetric input rates to search}

\item{vo_rng}{is the range of volumetric input rates to search}

\item{tp}{is the turbulent deposition probability}

\item{lp}{is the laminar deposition probability}

\item{it}{is the number of iterations to perform at each point}

\item{n}{is the number of years to simulate accumulation, an integer}
}
\value{
a list with elements (record, mean output rate)
}
\description{
For finding a needle in a haystack.
Uses backfill() to simulate accumulation.
}
\seealso{
backfill
}
