% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(probs)
}
\arguments{
\item{probs}{is a numeric vector of probabilities}
}
\value{
normalized vector of `probs`
}
\description{
normalize a vector of probabilities
}
