% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent_cdf}
\alias{percent_cdf}
\title{percent_cdf}
\usage{
percent_cdf(vec, ln = 100)
}
\arguments{
\item{vec}{is a numeric vector}

\item{ln}{is an integer}
}
\value{
a cdf length `ln ` of `vec`
}
\description{
computes the cdf of `vec`
coerces to length `ln`
}
