% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{plot_pt}
\alias{plot_pt}
\title{plot point}
\usage{
plot_pt(pt, mag = 1, so = nodes)
}
\arguments{
\item{pt}{is a coordinate pair}

\item{mag}{is a magnification factor (numeric)}

\item{so}{is a spatial object to plot the point upon}
}
\value{
a plot of `so` centered on `pt` at zoom factor `mag`
}
\description{
produces a plot centered around point `pt` zoomed in at factor `mag`
}
