% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{plot_weight}
\alias{plot_weight}
\title{plot_weight}
\usage{
plot_weight(num, den, by, bin_no = 10, labs)
}
\arguments{
\item{num}{is the numeric vector to use in the numerator of the weight}

\item{den}{is the numeric vector to use in the denomintor of the weight}

\item{by}{is the numeric vector to take the cdf of `num` and `dem` with respect to}

\item{bin_no}{is the number of bins to use in `bin_weights()`}

\item{labs}{is a character vector length 2 (xlab, ylab)}
}
\value{
a plot of the weighting function using `weight_by()`
}
\description{
prints a plot of the weighting function
produced using `bin_weights(num, dem, by, bin_no)`
}
\seealso{
weight_by
}
