% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pmf_by_rank}
\alias{pmf_by_rank}
\title{pmf by rank}
\usage{
pmf_by_rank(ranks, pmfs = char_pmfs)
}
\arguments{
\item{ranks}{is a vector of ranks (element of rank list)}

\item{pmfs}{is the matrix of charcoal pmfs}
}
\value{
the matrix of pmfs of rank `ranks`
}
\description{
wrapper for lapply, grabs pmfs with rank in `ranks` from main table
}
