% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{present_topology}
\alias{present_topology}
\title{present_topology}
\usage{
present_topology(ar, xs, ys)
}
\arguments{
\item{ar}{is a topology array}

\item{xs}{is a vector of values for the x-axis of the array}

\item{ys}{is a vector of values for the y-axis of the array}
}
\value{
prints a visualization of the topology array
}
\description{
prints a visualization of the topology array
}
