% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inher.R
\name{probcomb}
\alias{probcomb}
\title{Sums probability age value pairs into a single pmf}
\usage{
probcomb(x, y)
}
\arguments{
\item{x}{is a matrix of probability and age pairs}

\item{y}{is a numeric vector of years}
}
\value{
a numeric vector of length(years) with value of sum probability age is less than
   year i in y and greater than year i-1
}
\description{
Given a matrix of probability and age value pairs (x), and a vector of years (y),
returns a pmfs of length(y)
}
