% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rackl}
\alias{rackl}
\title{rack list}
\usage{
rackl(ls)
}
\arguments{
\item{ls}{is a list of data.tables of equal size}
}
\value{
data.tables in `ls` colbound into a matrix
}
\description{
turn a list of data.tables of equal size into a matrix via cbind
}
