% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ragged_bin}
\alias{ragged_bin}
\title{ragged_bin}
\usage{
ragged_bin(ln, bins = 10)
}
\arguments{
\item{ln}{is an integer specifying length}

\item{bins}{is an integer specifying number of bins}
}
\value{
a list length `bins` of index numbers
}
\description{
from indexes one to length `ln`
return a list of index numbers length `bins`
if `ln` is not equally divisible by `bins`
the last bin is smaller
}
