% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_list}
\alias{rank_list}
\title{Rank List}
\usage{
rank_list(ftype, dat = charcoal, min_count = 3)
}
\arguments{
\item{ftype}{is a character vector specifying facies type}

\item{dat}{is the data.table of charcoal site data}

\item{min_count}{is the minimum number of obs per site (1 is all sites)}
}
\value{
a list of sorted ranks at sites with obs > `min_count`
}
\description{
Utility for returning sorted ranks of obs at sites above a minimum count.
}
