% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{rater}
\alias{rater}
\title{rater}
\usage{
rater(so, gi, go, ti_scl, to_scl, ks_scl, dp_scl, m_scl, type, sa = sa_nodes)
}
\arguments{
\item{so}{is a spatial object (creek nodes)}

\item{gi}{is a global temporal input rate}

\item{go}{is a global temporal output rate}

\item{ti_scl}{is the range of temporal input scales to search}

\item{to_scl}{is the range of temporal output delivery scales to search}

\item{ks_scl}{is the range of temporal output streampower scales to search}

\item{dp_scl}{is the range of deposit probability scales to search}

\item{m_scl}{is the range of distance probability scales to search}

\item{type}{is a string indicated the type of output weight to apply}

\item{sa_nodes}{is the spatial object of sampled nodes}
}
\value{
spatial object with rates calculated
}
\description{
Derives variables related to temporal input and output rates.
The average rate per node is the global rate divided by the number of sampled nodes in sa_nodes.
Types of `type` include c('ks', 'dp', 'mixed'):
Streampower coefficient = 'ks'.
Inverse delivery probability = 'dp'.
Both = 'mixed'.
}
\seealso{
n_dimensional_search
}
