% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{rater1}
\alias{rater1}
\title{rater1}
\usage{
rater1(so, gi, go, sa = sa_nodes)
}
\arguments{
\item{so}{is a spatial object (creek nodes)}

\item{gi}{is a global temporal input rate}

\item{go}{is a global temporal output rate}

\item{sa_nodes}{is the spatial object of sampled nodes}
}
\value{
spatial object with rates calculated
}
\description{
Derives variables related to temporal input and output rates.
The average rate per node is the global rate divided by the number of sampled nodes in sa_nodes.
}
\seealso{
n_dimensional_search
}
