% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{rec_vol}
\alias{rec_vol}
\title{rec_vol}
\usage{
rec_vol(node, n, vi, vo)
}
\arguments{
\item{node}{is a spatial object (creek node)}

\item{n}{is an integer representing the number of years to accumulate}

\item{vi}{is the volume input rate, a float}

\item{vo}{is the volume output rate, a float}

\item{return}{matrix of volumes and levels}
}
\description{
function to record volumes
}
