% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{record_volume}
\alias{record_volume}
\title{record_volume}
\usage{
record_volume(node, n, vi, vo, arrivals = 0)
}
\arguments{
\item{node}{is a row from a spatial dataframe (creek nodes)}

\item{n}{is the number of years to simulate accumulation}

\item{vi}{is the volumetric input rate}

\item{vo}{is the volumetric output rate}

\item{arrivals}{is a dataframe describing upstream arrivals}
}
\value{
a list with elements c(accumulation, departures)
}
\description{
Record acculumation at a node.  Takes arrivals and emits departures.
Nodes are linked, with deposits departing downstream.
}
\seealso{
fit_volumes1
}
