% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{recorder}
\alias{recorder}
\title{recorder}
\usage{
recorder(n, ri, ro, ia)
}
\arguments{
\item{n}{is the number of accumulated arrivals}

\item{ri}{is the rate in, a float}

\item{ro}{is the rate out, a float}

\item{ia}{is a vector of inherited ages, integers}
}
\value{
deposit ages of accumulated arrivals, a numeric vector
}
\description{
models arrivals and removals from a reservoir
}
