% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{recorder1}
\alias{recorder1}
\title{recorder1}
\usage{
recorder1(n, ri, ro, ia, mn)
}
\arguments{
\item{n}{is the number of accumulated arrivals}

\item{ri}{is the rate in, a float}

\item{ro}{is the rate out, a float}

\item{ia}{is a vector of inherited ages, integers}

\item{mn}{is a minimum age}
}
\value{
deposit ages of accumulated arrivals, a numeric vector
}
\description{
models arrivals and removals from a reservoir
adds a minimum age
}
