% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{search_topology}
\alias{search_topology}
\title{search_topology}
\usage{
search_topology(
  so,
  x_range,
  y_range,
  rec = 0,
  bt = 10,
  it = 20,
  n = 10000,
  test = "ks",
  by = "vol",
  grid = 100,
  d_scl = 1,
  w_scl = 3,
  linked = FALSE,
  backfill = TRUE
)
}
\arguments{
\item{so}{is a spatial object (creek nodes)}

\item{x_range}{is a vector of values for the x-axis of the preference array}

\item{y_range}{is a vector of values for the y-axis of the preference array}

\item{rec}{is a data.frame of test values with cols c(x, y, ks, kp)}

\item{bt}{is the number of times to simulate accumulation over `so`}

\item{it}{is the number of times to simulate accumulation per node of `so`}

\item{n}{is the number of years to accumulate deposits}

\item{test}{is a string naming the test type c('ks', 'kp)}

\item{by}{is a string naming the test varaible c('vol', 'lvl')}

\item{grid}{is the number of intervals to divide x and y ranges into}

\item{d_scl}{is the scale value for weight_grid()}

\item{w_scl}{is the scale value for weight_array()}
}
\value{
a record of test value results at randomly generated coords in `x_range` and `y_range`
}
\description{
search_topology
}
