% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{select_rate_pairs}
\alias{select_rate_pairs}
\title{select_rate_pairs}
\usage{
select_rate_pairs(ar, xs, ys)
}
\arguments{
\item{ar}{is the preference matrix (a pdf superimposed on a grid)}

\item{xs}{is the vector of values along the x-axis of `ar`}

\item{ys}{is the vector of values along the y-axis of `ar`}
}
\value{
a pair of floats c(x, y)
}
\description{
function to return rate pairs given a preference matrix
}
