% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{snap}
\alias{snap}
\title{snap point}
\usage{
snap(pt, line = nodepath)
}
\arguments{
\item{pt}{is a coordinate pair}

\item{line}{is a matrix with rows (x,y) specifying a line path}
}
\value{
point on `line` closest in position to `pt`
}
\description{
finds the nearest point on a line path to a given point.
does not interpolate betweens points on line to find a local minimum.
}
