% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{snap_to_node}
\alias{snap_to_node}
\title{snap to node}
\usage{
snap_to_node(dist, outlet, ids)
}
\arguments{
\item{dist}{is a vector of outlet distances to place on the channel}

\item{outlet}{is a vector of outlet distances from the channel node network}

\item{ids}{is a vector of ids from the channel node network}
}
\value{
vector of ids from `ids` nearest to distances in `dist`
}
\description{
return id of node nearest to given outlet distance
}
