% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_tab}
\alias{stat_tab}
\title{stat_tab}
\usage{
stat_tab(mods, type = "lm")
}
\arguments{
\item{mods}{is a list of data.frames to fit using lm}
}
\value{
table of lm statistics from `mods`
}
\description{
print table of lm statistics from a list of data.frames
in each df, first var is dependent, other vars independent
stats include adj.r.squared, AIC and BIC
}
