% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inher.R
\name{sum_pdfs}
\alias{sum_pdfs}
\title{sum_pdfs add pdfs together and returns a normalized sum}
\usage{
sum_pdfs(lis, len)
}
\arguments{
\item{lis}{is a list of numeric pdfs}

\item{len}{is an integer representing the length of pdfs in lis}
}
\value{
a summed and normalized pdf
}
\description{
Given a list of numeric pdfs (lis) and an integer representing the length
of pdfs in lis (len), returns a pdf representing the normalized sum of pdfs
in lis
}
