% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_pmfs}
\alias{sum_pmfs}
\title{sum_pmfs add pmfs together and returns a normalized sum}
\usage{
sum_pmfs(lis, len)
}
\arguments{
\item{lis}{is a list of numeric pmfs}

\item{len}{is an integer representing the length of pmfs in lis}
}
\value{
a summed and normalized pmf
}
\description{
Given a list of numeric pmfs (lis) and an integer representing the length
of pmfs in lis (len), returns a pmf representing the normalized sum of pmfs
in lis
}
