% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{take_transect}
\alias{take_transect}
\title{take transect}
\usage{
take_transect(pt, bear, rad)
}
\arguments{
\item{pt}{is a coordinate pair (centerpoint of transect)}

\item{bear}{is a bearing in degrees (angle of transect)}

\item{rad}{is a radius specifying (width of transect)}
}
\value{
matrix of coordinate pairs along transect using inc_path()
}
\description{
given a point pt, a bearing bear and a radius rad
take a transect centered on point pt, along bearing bear with radius rad
}
\seealso{
tri_length

inc_path
}
