% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluvial.R
\name{test_fit}
\alias{test_fit}
\title{test_fit}
\usage{
test_fit(
  fits,
  ad_a = 340,
  ad_b = 280,
  ch_a = 200,
  ch_b = 150,
  kp_a = 0.12,
  kp_b = 0.09,
  ks_a = 0.09,
  ks_b = 0.05
)
}
\arguments{
\item{fits}{Vector of fits from gof().}

\item{ad_a}{Less stringent threshold for Anderson-Darling test.}

\item{ad_b}{More stringent threshold for Anderson-Darling test.}

\item{ch_a}{Less stringent threshold for Chi-squared test.}

\item{ch_b}{More stringent threshold for Chi-squared test.}

\item{kp_a}{Less stringent threshold for Kuiper test.}

\item{kp_b}{More stringent threshold for Kuiper test.}

\item{ks_a}{Less stringent threshold for Kolmogorov-Smirnov test.}

\item{ks_b}{More stringent threshold for Kolmogorov-Smirnov test.}
}
\value{
Vector of hit % = c(ad_a, ch_a, kp_a, ks_a, ad_b, ch_b, kp_b, ks_b)
}
\description{
Hit statistics on how often a vector of fits passes a set of thresholds.
}
