% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timer}
\alias{timer}
\title{timer}
\usage{
timer(exp, msg = "time elapsed: ")
}
\arguments{
\item{exp}{is an expression}

\item{msg}{is a string allowing a customizable message for output}
}
\value{
prints the time elapsed after expression computes with `msg`
}
\description{
wrapper for expressions to measure time to compute
}
