% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inher.R
\name{to_cdf}
\alias{to_cdf}
\title{to_cdf converts numeric pdf vector to cdf of equal length}
\usage{
to_cdf(vec)
}
\arguments{
\item{vec}{is a numeric pdf}
}
\value{
a numeric vector of the cdf
}
\description{
Given a numeric pdf, returns a numeric vector of the cdf.
}
