% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inher.R
\name{to_exceed}
\alias{to_exceed}
\title{to_exceed converts numeric pmf vectors to log10 exceedance distributions of equal length}
\usage{
to_exceed(vec)
}
\arguments{
\item{vec}{is a numeric pmf}
}
\value{
a numeric vector of the log10 exceedance distribution
}
\description{
Given a numeric pmf, returns a numeric vector of the log10 exceedance distribution
}
\seealso{
to_cdf
}
