% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_mat}
\alias{to_mat}
\title{to matrix}
\usage{
to_mat(ls)
}
\arguments{
\item{ls}{is a list of data.tables}
}
\value{
a matrix merged by cbind
}
\description{
from list of data.tables to matrix using cbind
}
