% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inher.R
\name{to_pdf}
\alias{to_pdf}
\title{CONVERT VECTOR TO PDF}
\usage{
to_pdf(cdf)
}
\arguments{
\item{cdf}{is a numeric vector (a cdf)}
}
\value{
the pdf derived from \code{cdf} as a numeric vector
}
\description{
Given the cdf as a numeric vector, return the derived pdf as a numeric vector of equal length.
}
