% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_pmf}
\alias{to_pmf}
\title{CONVERT VECTOR TO pmf}
\usage{
to_pmf(cdf)
}
\arguments{
\item{cdf}{is a numeric vector (a cdf)}
}
\value{
the pmf derived from `cdf` as a numeric vector
}
\description{
Given the cdf as a numeric vector, return the derived pmf as a numeric vector of equal length.
}
