% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tracker}
\alias{tracker}
\title{tracker}
\usage{
tracker(i, n, begin)
}
\arguments{
\item{i}{is an integer specifying the index position}

\item{n}{is an integer specifying the number of interations in the loop}

\item{begin}{is a time object from a call to Sys.time()}
}
\value{
prints percent complete and time remaining estimate to console
}
\description{
prints percent complete and time remaining estimate to console
}
