% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{tri_length}
\alias{tri_length}
\title{triangle length}
\usage{
tri_length(pt, bear, dist, north = TRUE)
}
\arguments{
\item{pt}{is a coordinate pair}

\item{bear}{is a bearing in degrees}

\item{dist}{is a numeric distance (hypotenuse length)}

\item{north}{is a boolean indicating north or south}
}
\value{
coords of end point along hypotenuse
}
\description{
Given a coordinate pair, and the length of the hypotenuse
of a right triangle, returns the coordinate pair at the
other end of the segment.
}
