% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trunc_0}
\alias{trunc_0}
\title{truncate at zero}
\usage{
trunc_0(pmf, index)
}
\arguments{
\item{pmf}{is a numeric vector of probabilites}

\item{index}{is a numeric vector values associates with `pmf`}
}
\value{
`pmf` with probs at vals below zero set to zero
}
\description{
truncate a pmf at index values below zero
}
