% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trunc_list}
\alias{trunc_list}
\title{trunc list}
\usage{
trunc_list(pmfs, index)
}
\arguments{
\item{pmfs}{is a list of numeric vectors of probabilities that sum to 1}

\item{index}{is a numeric vector of values associated with probs in `pmfs`}
}
\value{
truncated and renormalized pmfs in list
}
\description{
truncate pmfs in list at values below zero and renormalize
}
