% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trunc_norm}
\alias{trunc_norm}
\title{truncate & normalize}
\usage{
trunc_norm(pmf, index)
}
\arguments{
\item{pmf}{is a numeric vector of probabilities}

\item{index}{is a numeric vector of values associated with `pmf`}
}
\value{
`pmf` truncated at `index` values below zero and renormalized
}
\description{
truncates `pmf` at `index` values below zero and renormalizes
}
