% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{update_search_array}
\alias{update_search_array}
\title{update_search_array}
\usage{
update_search_array(rec, xs, ys, ar, test, by, d_scl = 1, w_scl = 0.9)
}
\arguments{
\item{rec}{is a data.frame of test values with cols c(x, y, ks, kp)}

\item{xs}{is a vector of values for the x-axis of topology array}

\item{ys}{is a vector of values for the y-axis of topology array}

\item{ar}{is a preference topology to update}

\item{test}{is a string naming the test type c('ks', 'kp)}

\item{by}{is a string naming the test varaible c('vol', 'lvl')}

\item{d_scl}{is the scale value for weight_grid()}

\item{w_scl}{is the scale value for weight_array()}
}
\value{
a preference topology array
}
\description{
Returns preference topology given a record of test values.
Types for `test`:
Kolmogorov-Smirnov statistic = 'ks'
Kuiper statistic = 'kp'
Types for `by`:
Unit cross-sectional area in m2 = 'vol'
Average cross-sectional depth per node = 'lvl'
}
