% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{val_by_cdf}
\alias{val_by_cdf}
\title{val_by_cdf}
\usage{
val_by_cdf(vals, ln = 100)
}
\arguments{
\item{ln}{is an integer}

\item{val}{is a numeric vector}
}
\value{
values of `val` at cdf of `val`
}
\description{
values at position of cdf of `val`
}
