% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{voluminous}
\alias{voluminous}
\title{voluminous}
\usage{
voluminous(nodes, n, vi, vo, arrivals)
}
\arguments{
\item{nodes}{is a spatial object (creek nodes)}

\item{n}{is the number of years to simulate accumulation, an integer}

\item{vi}{is the volumetric input rate}

\item{vo}{is the volumetric output rate}

\item{arrivals}{is the record of arrivals, a data.frame with cols (t, r, vol, lvl)}
}
\value{
a list with elements c(volumes, levels, arrivals)
}
\description{
Wrapper for record_volume.
}
