% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{voluminous1}
\alias{voluminous1}
\title{voluminous1}
\usage{
voluminous1(nodes, n, vi, vo, b_scl)
}
\arguments{
\item{nodes}{is a spatial object (creek nodes)}

\item{n}{is the number of years to simulate accumulation, an integer}

\item{vi}{is the volumetric input rate}

\item{vo}{is the volumetric output rate}

\item{b_scl}{is the proportion of volume that settles out of deposits at obstructions}
}
\value{
a list with elements c(volumes, levels, arrivals)
}
\description{
Simulates accumulation record, using a linked-bucket model
with backfilling based on average elevation of bank deposits.
}
\seealso{
eventer
}
