% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{voluminous2}
\alias{voluminous2}
\title{voluminous2}
\usage{
voluminous2(nodes, n, vi, vo, tp = 0.1, lp = 0.5)
}
\arguments{
\item{nodes}{is a spatial object (creek nodes)}

\item{n}{is the number of years to simulate accumulation, an integer}

\item{vi}{is the volumetric input rate}

\item{vo}{is the volumetric output rate}

\item{tp}{is the turbulent deposition probability}

\item{lp}{is the laminar deposition probability}
}
\value{
a list with elements c(volumes, levels, arrivals)
}
\description{
Simulates accumulation record, using a linked-bucket model
with backfilling based on average elevation of bank deposits.
}
\seealso{
eventer
}
