% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{weight_array}
\alias{weight_array}
\title{weight_array}
\usage{
weight_array(ar, scl = 1)
}
\arguments{
\item{ar}{is a topology matrix from weight_grid()}

\item{scl}{is the scale parameter}
}
\value{
a preference matrix
}
\description{
Returns preference matrix given topology matrix.
The scale parameter deflates preference weighting from zero to one,
inflates preference weighting above one,
and crashes below zero.
}
\seealso{
weight_grid
}
