% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{weight_by}
\alias{weight_by}
\title{weight_by}
\usage{
weight_by(num, den, by, bin_no = 10)
}
\arguments{
\item{num}{is the numeric vector to use in the numerator of the weight}

\item{den}{is the numeric vector to use in the denomintor of the weight}

\item{by}{is the numeric vector to take the cdf of `num` and `dem` with respect to}

\item{bin_no}{is the number of bins to use in `bin_weights()`}
}
\value{
a numeric vector of weighting function
}
\description{
generates a weighting function
produced using `bin_weights(num, dem, by, bin_no)`
}
\seealso{
bin_weights
}
