% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{weight_by_delprob}
\alias{weight_by_delprob}
\title{weight_by_delprob}
\usage{
weight_by_delprob(delprob, weight = df_wt)
}
\arguments{
\item{delprob}{is a vector of delivery probabilities, f64}

\item{weight}{is a matrix with cols c(cdf, delprob)}
}
\value{
vector of optimized delivery probabilities
}
\description{
converts MB delivery probabilities to Rose-Lancaster optimized delivery probabilites
}
