% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{zoom}
\alias{zoom}
\title{zoom}
\usage{
zoom(x, d, z = 2)
}
\arguments{
\item{x}{is a spatial object}

\item{d}{is coordinate pair (centerpoint)}

\item{z}{is numeric zoom factor}
}
\value{
spatial object scaled to new extent by zoom factor
}
\description{
Zoom in on a raster like a microscope
draws a new extent m(x-,x+,y-,y+)
d(x,y) is centerpoint of raster
z is zoom factor scaled by ((x+ - x-)|(y+ - y-))/(2 * z)
}
