% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inher.R
\name{convo}
\alias{convo}
\title{derived pmf by subtraction}
\usage{
convo(x, y, index)
}
\arguments{
\item{x}{is a numeric vector (younger pdf)}

\item{y}{is a numeric vector (older pdf)}

\item{index}{is a numeric vector (years)}
}
\value{
a length(index) numeric vector of the convolved distribution of y-x
}
\description{
Given two pdfs and an age vector, returns the derived pdf of the convolution y-x.
Lengths of x,y and index must be equal.
}
\seealso{
convo_minus
}
