% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convo}
\alias{convo}
\title{derived pmf by subtraction}
\usage{
convo(x, y, index)
}
\arguments{
\item{x}{is a numeric vector (older pmf)}

\item{y}{is a numeric vector (younger pmf)}

\item{index}{is a numeric vector (years)}
}
\value{
a length(index) numeric vector of the convolved distribution of x-y
}
\description{
Given two pmfs and an age vector, returns the derived pmf of the convolution x-y.
Lengths of x,y and index must be equal.
}
\seealso{
convo_minus
}
