% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convo_lis}
\alias{convo_lis}
\title{convolve list}
\usage{
convo_lis(lis, pmfs = char_pmfs)
}
\arguments{
\item{lis}{is a list of integer ranks}

\item{pmfs}{is a matrix of pmfs with rownames(vals) (charcoal ages)}
}
\value{
a list of convolved `pmfs` with ranks in `lis`
}
\description{
returns a list of convolved `pmfs` with ranks in `lis`
}
