% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convo_ranks}
\alias{convo_ranks}
\title{convolve vector of ranks}
\usage{
convo_ranks(vec, pmfs = char_pmfs)
}
\arguments{
\item{vec}{is a vector of sorted numeric ranks}

\item{pmfs}{is a matrix of pmfs with cols of obs by rank and rows of probs}
}
\value{
a matrix of pmfs convolved by subtracting pmf of lowest rank from pmfs
  with ranks in `vec`
}
\description{
subtracts the pmf with first rank in `vec` from pmfs with ranks in `vec`
via convolution
}
