% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discrete_bin}
\alias{discrete_bin}
\title{discrete bin}
\usage{
discrete_bin(vals, bin_len)
}
\arguments{
\item{vals}{is the vector of values to bin}

\item{bin_len}{is the size of bin}
}
\value{
a list that splits vals into elements of length `bin_len`
}
\description{
split a vector of values into discrete bins of roughly equal size
}
