% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_bins}
\alias{fit_bins}
\title{fit bins}
\usage{
fit_bins(bins)
}
\arguments{
\item{bins}{is the binned data to fit}
}
\value{
exponential fits for each bin, gof stats and CIs
}
\description{
fit an exponential distribution to binned data
}
