% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rack}
\alias{rack}
\title{Rack List}
\usage{
rack(ls)
}
\arguments{
\item{ls}{is a list with elements of equal length}
}
\value{
elements of `ls` column bound into a matrix
}
\description{
Rack list with elements of equal length into a matrix using cbind
}
