% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_tab}
\alias{stat_tab}
\title{stat_tab}
\usage{
stat_tab(df, type = "lm")
}
\arguments{
\item{df}{is a data.frame object}
}
\value{
table of lm statistics from `df`
}
\description{
in `df`, first var is dependent, other vars independent
gets all formula combinations
returns a table of model stats
adj.r.squared, AIC and BIC
}
