% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_cdf}
\alias{to_cdf}
\title{to_cdf converts numeric pmf vector to cdf of equal length}
\usage{
to_cdf(vec)
}
\arguments{
\item{vec}{is a numeric pmf}
}
\value{
a numeric vector of the cdf
}
\description{
Given a numeric pmf, returns a numeric vector of the cdf.
}
